[ENG-5751] remove old keen logic (#11082)
* ENG-5751 remove old keen logic
* add migration
* remove legacy keen from metrics .js (maybe another keen logic is also needed to be removed in the files)
* remove dashboard keen (it is not used in any views)
* ENG-5751 code updates
* ENG-5751 remove keen dependency from poetry
* ENG-5751 code updates
* code updates
* ENG-5751 resolve PR comments
* ENG-5751 add default '' analytics_key
* ENG-5751 remove base_analytics.html
* add keen-dashboards.css
* fix/ENG-5751 it seems 'keen' is used in base mako files and keen js rely on it and it is not needed to be deleted
* fix/ENG-5751 resolve PR comments (not confident about 'getUniqueId' implementation https://stackoverflow.com/questions/18230217/javascript-generate-a-random-number-within-a-range-using-crypto-getrandomvalues )
* remove keen-tracking
* update getUniqueId
* fix/ENG-5751 PR comments related updates
* fix/ENG-5751 PR comments related updates
* remove geoip from deps and location code from app
* PR comment related updates
https://github.com/CenterForOpenScience/osf.io/pull/10992#discussion_r2021990050
* fix migrations conflicts
* a few tidy-ups:
 * Relocate code changes in api.nodes.serializers to be more compact
   when squished.
 * Really delete `cumulative_plos_metrics.py`. Per Eric's comments on
   Slack, we no longer need this.
 * Remove `cumulative_plos_metrics.py`-related config from
   website.settings.defaults.py
 * Add attributive comments for borrowed code in metrics library.

---------
