+++ admin/metrics/views.py
@@ -2,7 +2,6 @@ from django.views.generic import TemplateView
 from django.contrib.auth.mixins import PermissionRequiredMixin
 
 from admin.base import settings
-from admin.base.settings import KEEN_CREDENTIALS
 
 
 class MetricsView(PermissionRequiredMixin, TemplateView):
@@ -11,7 +10,6 @@ class MetricsView(PermissionRequiredMixin, TemplateView):
     raise_exception = True
 
     def get_context_data(self, **kwargs):
-        kwargs.update(KEEN_CREDENTIALS.copy())
         api_report_url = f'{settings.API_DOMAIN}_/metrics/reports/'
         kwargs.update({'metrics_url': api_report_url})
         return super().get_context_data(**kwargs)
