+++ api_tests/nodes/serializers/test_serializers.py
@@ -48,7 +48,6 @@ class TestNodeSerializer:
         assert attributes['registration'] == node.is_registration
         assert attributes['fork'] == node.is_fork
         assert attributes['collection'] == node.is_collection
-        assert attributes['analytics_key'] == node.keenio_read_key
         assert attributes['wiki_enabled'] == node.has_addon('wiki')
 
         # Relationships
