+++ package.json
@@ -33,9 +33,7 @@
     "js-cookie": "3.0.1",
     "js-md5": "^0.7.3",
     "jstimezonedetect": "^1.0.6",
-    "keen-analysis": "^1.1.0",
     "keen-dataviz": "^1.0.2",
-    "keen-tracking": "1.0.3",
     "knockout": "~3.4.2",
     "knockout.validation": "^2.0.2",
     "less": "^4.1.2",
