+++ website/settings/local-ci.py
@@ -82,23 +82,8 @@ VARNISH_SERVERS = ['http://localhost:8080']
 # if ENABLE_VARNISH isn't set in python read it from the env var and set it
 locals().setdefault('ENABLE_VARNISH', os.environ.get('ENABLE_VARNISH') == 'True')
 
-KEEN = {
-    'public': {
-        'project_id': '123456789abcdef101112131415161718191a1b1c1d1e1f20212223242526272',
-        'master_key': '123456789abcdef101112131415161718191a1b1c1d1e1f20212223242526272',
-        'write_key': '123456789abcdef101112131415161718191a1b1c1d1e1f20212223242526272',
-        'read_key': '123456789abcdef101112131415161718191a1b1c1d1e1f20212223242526272',
-    },
-    'private': {
-        'project_id': '123456789abcdef101112131415161718191a1b1c1d1e1f20212223242526272',
-        'write_key': '123456789abcdef101112131415161718191a1b1c1d1e1f20212223242526272',
-        'read_key': '123456789abcdef101112131415161718191a1b1c1d1e1f20212223242526272',
-    },
-}
-
 NEW_AND_NOTEWORTHY_LINKS_NODE = 'helloo'
 POPULAR_LINKS_NODE = 'hiyah'
-POPULAR_LINKS_REGISTRATIONS = 'woooo'
 
 logging.getLogger('celery.app.trace').setLevel(logging.FATAL)
