+++ website/templates/base.mako
@@ -258,7 +258,6 @@
                     entryPoint: ${ user_entry_point | sjson, n },
                     institutions: ${ user_institutions | sjson, n},
                     emailsToAdd: ${ user_email_verifications | sjson, n },
-                    anon: ${ anon | sjson, n },
                 },
                 maintenance: ${ maintenance | sjson, n},
                 analyticsMeta: {
@@ -271,23 +270,6 @@
             });
         </script>
 
-        % if keen['public']['project_id']:
-            <script>
-                window.contextVars = $.extend(true, {}, window.contextVars, {
-                    keen: {
-                        public: {
-                            projectId: ${ keen['public']['project_id'] | sjson, n },
-                            writeKey: ${ keen['public']['write_key'] | sjson, n },
-                        },
-                        private: {
-                            projectId: ${ keen['private']['project_id'] | sjson, n },
-                            writeKey: ${ keen['private']['write_key'] | sjson, n },
-                        },
-                    },
-                });
-            </script>
-        % endif
-
         % if settings.DATACITE_TRACKER_REPO_ID:
             <script>
                 window.contextVars = $.extend(true, {}, window.contextVars, {
@@ -295,7 +277,6 @@
                 });
             </script>
         % endif
-
 
         ${self.javascript_bottom()}
     </body>
